+++ docs/source/gallery/examples/geochem/indexes_selectors.py
@@ -8,9 +8,9 @@ import pandas as pd
 
 pd.set_option("precision", 3)  # smaller outputs
 ########################################################################################
-from pyrolite.util.synthetic import test_df
+from pyrolite.util.synthetic import normal_frame
 
-df = test_df(cols=["CaO", "MgO", "SiO2", "FeO", "Mn", "Ti", "La", "Lu", "Mg/Fe"])
+df = normal_frame(columns=["CaO", "MgO", "SiO2", "FeO", "Mn", "Ti", "La", "Lu", "Mg/Fe"])
 ########################################################################################
 
 df.head(2).pyrochem.oxides