+++ test/comp/comp_pyrocomp.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 import pyrolite.comp
-from pyrolite.util.synthetic import test_df
+from pyrolite.util.synthetic import normal_frame
 from pyrolite.geochem.ind import REE
 
 np.random.seed(81)
@@ -12,9 +12,9 @@ class TestPyroComp(unittest.TestCase):
         self.cols = ["MgO", "SiO2", "CaO"]
 
         # can run into interesting singular matrix errors with bivariate random data
-        self.tridf = test_df(cols=self.cols, index_length=100)
+        self.tridf = normal_frame(columns=self.cols, size=100)
         self.bidf = self.tridf.loc[:, self.cols[:2]]
-        self.multidf = test_df(cols=REE(), index_length=100)
+        self.multidf = normal_frame(columns=REE(), size=100)
 
     def test_renormalise_default(self):
         df = self.bidf.copy(deep=True) * 100  # copy df