+++ test/util/skl/util_skl_select.py
@@ -1,6 +1,6 @@
 import unittest
 import numpy as np
-from pyrolite.util.synthetic import test_df
+from pyrolite.util.synthetic import normal_frame
 from pyrolite.comp.codata import close
 
 try:
@@ -26,7 +26,7 @@ class TestSelectors(unittest.TestCase):
     """Checks the default config for scikit-learn selector classes."""
 
     def setUp(self):
-        self.df = test_df().apply(close, axis=1)
+        self.df = normal_frame().apply(close, axis=1)
 
     def test_ColumnSelector(self):
         """Test the ColumnSelector transfomer."""